backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# Redis for pub/sub
redis==5.0.1
aioredis==2.0.1

# WebSocket support
websockets==12.0

# HTTP client
httpx==0.25.2

# Data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Code quality
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Monitoring and logging
structlog==23.2.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

